move all test-only functionality to tetinterfaces.go
