+++ ledger/acctupdates_test.go
@@ -267,7 +267,8 @@ func (au *accountUpdates) allBalances(rnd basics.Round) (bals map[basics.Address
 		if err != nil {
 			return err
 		}
-		bals, err0 = arw.AccountsAllTest()
+		testArw := arw.(store.TestAccountsReaderExt)
+		bals, err0 = testArw.AccountsAllTest()
 		return err0
 	})
 	if err != nil {
@@ -1005,11 +1006,12 @@ func TestListCreatables(t *testing.T) {
 		proto := config.Consensus[protocol.ConsensusCurrentVersion]
 
 		accts := make(map[basics.Address]basics.AccountData)
-		_ = tx.AccountsInitTest(t, accts, protocol.ConsensusCurrentVersion)
+		testTx := tx.(store.TransactionTestScope)
+		_ = testTx.AccountsInitTest(t, accts, protocol.ConsensusCurrentVersion)
 		require.NoError(t, err)
 
 		au := &accountUpdates{}
-		au.accountsq, err = tx.MakeAccountsOptimizedReader()
+		au.accountsq, err = testTx.MakeAccountsOptimizedReader()
 		require.NoError(t, err)
 
 		// ******* All results are obtained from the cache. Empty database *******
